k12_platform/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.12.0
jinja2==3.1.3
aiofiles==23.2.1
pillow==10.2.0
